I just went thought and made some miner changes.
added @staticmethod to static class methods that are static
cleaned up the Neron setBias to be more clear and handle all cases
renamed randomizWeights to randomizeWeights in the Neuron class
removed redundant parentheses

Added error checks to let us know in Driver.py if network_in has an error where testing_data or training_data is of size zero. That wouldn't have caused an issue before.

Added a controller.py and view.py so if we end up using MVC we will have files for those.